refactor(auth): route project-permission check through service layer (#4053)
Extract permission enforcement logic from a free function in auth/permissions.ts into a new PermissionsService class in app-layer/permissions/permissions.service.ts. The free function becomes a thin wrapper that instantiates the service, keeping all existing call sites unchanged.
